requirements_simple.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
requests==2.31.0
